Code spiltting for right side features
