+++ src/components/Table/ConnectPointsWrapper.tsx
@@ -1,0 +1,59 @@
+import React, { useRef, useState } from "react";
+import Xarrow from "react-xarrows";
+
+const connectPointStyle = {
+  position: "absolute",
+  width: 15,
+  height: 15,
+  borderRadius: "50%",
+  background: "black",
+};
+const connectPointOffset = {
+  left: { left: 0, top: "50%", transform: "translate(-50%, -50%)" },
+  right: { left: "100%", top: "50%", transform: "translate(-50%, -50%)" },
+  top: { left: "50%", top: 0, transform: "translate(-50%, -50%)" },
+  bottom: { left: "50%", top: "100%", transform: "translate(-50%, -50%)" },
+};
+
+const ConnectPointsWrapper = ({ boxId, handler, dragRef, boxRef }) => {
+  const ref1 = useRef();
+
+  const [position, setPosition] = useState({});
+  const [beingDragged, setBeingDragged] = useState(false);
+  return (
+    <React.Fragment>
+      <div
+        className="connectPoint"
+        style={{
+          ...connectPointStyle,
+          ...connectPointOffset[handler],
+          ...position,
+        }}
+        draggable
+        onMouseDown={(e) => e.stopPropagation()}
+        onDragStart={(e) => {
+          setBeingDragged(true);
+          e.dataTransfer.setData("arrow", boxId);
+        }}
+        onDrag={(e) => {
+          const { offsetTop, offsetLeft } = boxRef.current;
+          const { x, y } = dragRef.current.state;
+          setPosition({
+            position: "fixed",
+            left: e.clientX - x - offsetLeft,
+            top: e.clientY - y - offsetTop,
+            transform: "none",
+            opacity: 0,
+          });
+        }}
+        ref={ref1}
+        onDragEnd={(e) => {
+          setPosition({});
+          setBeingDragged(false);
+        }}
+      />
+      {beingDragged ? <Xarrow start={boxId} end={ref1} /> : null}
+    </React.Fragment>
+  );
+};
+export default ConnectPointsWrapper;
